Todo list project
This small Django project which named as Todo. In this we can store our daily tasks, we can change their status base on the progress . To make this kind of project  DTL(Django Template language is used) .If you are the one of users this project/app you are allowed to use the project to add the tasks  and to fetch your tasks etc.